Move AbstractSparseArray to SparseArrayInterface
