+++ NDTensors/src/lib/BlockSparseArrays/src/abstractblocksparsearray.jl
@@ -1,6 +1,7 @@
 using BlockArrays: BlockArrays, AbstractBlockArray
 
-# TODO: Delete this.
+# TODO: Delete this. This function was replaced
+# by `nstored` but is still used in `NDTensors`.
 function nonzero_keys end
 
 abstract type AbstractBlockSparseArray{T,N} <: AbstractBlockArray{T,N} end
