+++ NDTensors/src/lib/DiagonalArrays/src/DiagonalArrays.jl
@@ -3,9 +3,10 @@ include("defaults.jl")
 include("diaginterface.jl")
 include("diagindex.jl")
 include("diagindices.jl")
-include("diagonalarray.jl")
+include("abstractdiagonalarray.jl")
 include("sparsearrayinterface.jl")
 include("diagonalarraydiaginterface.jl")
+include("diagonalarray.jl")
 include("diagonalmatrix.jl")
 include("diagonalvector.jl")
 end
