+++ NDTensors/src/lib/DiagonalArrays/src/abstractdiagonalarray.jl
@@ -1,0 +1,3 @@
+using ..SparseArrayDOKs: AbstractSparseArray
+
+abstract type AbstractDiagonalArray{T,N} <: AbstractSparseArray{T,N} end
