--- NDTensors/src/lib/SparseArrayDOKs/src/convert.jl
@@ -1,1 +1,0 @@
-Base.convert(type::Type{<:AbstractSparseArray}, a::AbstractArray) = type(a)
