+++ NDTensors/src/lib/SparseArrayInterface/.JuliaFormatter.toml
@@ -1,0 +1,2 @@
+style = "blue"
+indent = 2
