+++ NDTensors/src/lib/SparseArrayInterface/src/SparseArrayInterface.jl
@@ -1,14 +1,21 @@
 module SparseArrayInterface
-include("densearray.jl")
-include("interface.jl")
-include("interface_optional.jl")
-include("indexing.jl")
-include("base.jl")
-include("map.jl")
-include("copyto.jl")
-include("broadcast.jl")
-include("conversion.jl")
-include("wrappers.jl")
-include("zero.jl")
-# include("SparseArrayInterfaceSparseArraysExt.jl")
+include("sparsearrayinterface/densearray.jl")
+include("sparsearrayinterface/interface.jl")
+include("sparsearrayinterface/interface_optional.jl")
+include("sparsearrayinterface/indexing.jl")
+include("sparsearrayinterface/base.jl")
+include("sparsearrayinterface/map.jl")
+include("sparsearrayinterface/copyto.jl")
+include("sparsearrayinterface/broadcast.jl")
+include("sparsearrayinterface/conversion.jl")
+include("sparsearrayinterface/wrappers.jl")
+include("sparsearrayinterface/zero.jl")
+include("abstractsparsearray/abstractsparsearray.jl")
+include("abstractsparsearray/sparsearrayinterface.jl")
+include("abstractsparsearray/base.jl")
+include("abstractsparsearray/broadcast.jl")
+include("abstractsparsearray/map.jl")
+include("abstractsparsearray/baseinterface.jl")
+include("abstractsparsearray/convert.jl")
+include("abstractsparsearray/SparseArrayDOKsSparseArraysExt.jl")
 end
