+++ NDTensors/src/lib/SparseArrayInterface/src/abstractsparsearray/convert.jl
@@ -1,0 +1,5 @@
+Base.convert(type::Type{<:AbstractSparseArray}, a::AbstractArray) = type(a)
+
+function (::Type{T})(a::T) where {T<:AbstractSparseArray}
+  return copy(a)
+end
