+++ NDTensors/src/lib/SparseArrayInterface/src/abstractsparsearray/sparsearrayinterface.jl
@@ -1,7 +1,7 @@
 using Dictionaries: set!
 using ..SparseArrayInterface: SparseArrayInterface
 
-SparseArrayInterface.storage(::AbstractSparseArray) = error("Not implemented")
+SparseArrayInterface.sparse_storage(::AbstractSparseArray) = error("Not implemented")
 
 # TODO: Move to `SparseArrayInterface`?
 getindex_zero_function(::AbstractSparseArray) = error("Not implemented")
@@ -9,7 +9,7 @@ getindex_zero_function(::AbstractSparseArray) = error("Not implemented")
 function SparseArrayInterface.index_to_storage_index(
   a::AbstractSparseArray{<:Any,N}, I::CartesianIndex{N}
 ) where {N}
-  !isassigned(SparseArrayInterface.storage(a), I) && return nothing
+  !isassigned(SparseArrayInterface.sparse_storage(a), I) && return nothing
   return I
 end
 
@@ -30,10 +30,6 @@ end
 function SparseArrayInterface.setindex_notstored!(
   a::AbstractSparseArray{<:Any,N}, value, I::CartesianIndex{N}
 ) where {N}
-  set!(SparseArrayInterface.storage(a), I, value)
+  set!(SparseArrayInterface.sparse_storage(a), I, value)
   return a
-end
-
-function SparseArrayInterface.empty_storage!(a::AbstractSparseArray)
-  return empty!(SparseArrayInterface.storage(a))
 end
