+++ NDTensors/src/lib/SparseArrayInterface/src/sparsearrayinterface/interface.jl
@@ -4,7 +4,7 @@
 # Minimal interface
 # Data structure of the stored (generally nonzero) values
 # nonzeros(a::AbstractArray) = a
-storage(a::AbstractArray) = a
+sparse_storage(a::AbstractArray) = a
 
 # Minimal interface
 # Map an index into the stored data to a CartesianIndex of the
