+++ NDTensors/src/lib/SparseArrayInterface/test/SparseArrayInterfaceTestUtils/SparseArrays.jl
@@ -29,7 +29,7 @@ function Base.similar(a::SparseArray, elt::Type, dims::Tuple{Vararg{Int}})
 end
 
 # Minimal interface
-SparseArrayInterface.storage(a::SparseArray) = a.data
+SparseArrayInterface.sparse_storage(a::SparseArray) = a.data
 function SparseArrayInterface.index_to_storage_index(
   a::SparseArray{<:Any,N}, I::CartesianIndex{N}
 ) where {N}
@@ -47,10 +47,11 @@ end
 
 # Empty the storage, helps with efficiency in `map!` to drop
 # zeros.
-function SparseArrayInterface.empty_storage!(a::SparseArray)
+function SparseArrayInterface.dropall!(a::SparseArray)
   empty!(a.data)
   empty!(a.index_to_dataindex)
-  return empty!(a.dataindex_to_index)
+  empty!(a.dataindex_to_index)
+  return a
 end
 
 # Broadcasting
